+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * @iov/lisk: Implement `LiskConnection.watchAccount`
 * @iov/lisk: Support search by address tag in `LiskConnection.searchTx`
+* @iov/lisk: Support search by minHeight/maxHeight in `LiskConnection.searchTx`
 * @iov/rise: Implement `RiseConnection.watchAccount`
 * @iov/rise: Support search by address tag in `RiseConnection.searchTx`
 * @iov/keycontrol: Fix parameter type of `UserProfile.createIdentity` to allow